refactor(pipeline): split Pipeline into focused mixin modules
Splits the ~900-line pipeline.py into six focused files while keeping
the public API (Pipeline, scrape, process_url, process_urls) identical.
No behavior changes — pure structural refactor (CAS-49).

- pipeline.py: spine only (~200 lines of init + public methods)
- pipeline_cache.py: cached selector replay and per-field staleness
- pipeline_extraction.py: fetch, clean, extract, downloads, verify
- pipeline_discovery.py: AI discovery, MCP escalation, JS actions
- pipeline_crawler.py: placeholder for CAS-52 crawl helpers
- pipeline_utils.py: stateless helpers, validation, display methods
- _pipeline_table.py: shared concurrent-progress table builder

Uses Python mixins so all methods retain full self access. Updates
yosoi/cli/progress.py and yosoi/core/AGENTS.md accordingly.
