+++ yosoi/core/pipeline_table.py
@@ -1,0 +1,33 @@
+"""Shared Rich table builder for concurrent URL progress display.
+
+Extracted from pipeline.py so it can be imported by both the pipeline spine
+and the CLI backwards-compat stub without creating a circular dependency.
+"""
+
+from __future__ import annotations
+
+import time
+
+from rich.table import Table
+
+_STATUS_STYLES: dict[str, tuple[str, bool]] = {
+    'Queued': ('dim', False),
+    'Running': ('bold yellow', True),
+    'Done': ('bold green', False),
+    'Skipped': ('dim', False),
+    'Failed': ('bold red', False),
+}
+
+
+def _build_concurrent_table(url_status: dict[str, tuple[str, float]]) -> Table:
+    """Build a Rich Table showing per-URL concurrent progress."""
+    table = Table(title='Concurrent Processing', expand=True)
+    table.add_column('#', style='dim', width=4)
+    table.add_column('URL', style='cyan', ratio=3)
+    table.add_column('Status', width=12)
+    table.add_column('Elapsed', style='dim', width=10)
+    for idx, (u, (status, value)) in enumerate(url_status.items(), 1):
+        style, is_running = _STATUS_STYLES.get(status, ('bold red', False))
+        elapsed_str = f'{time.monotonic() - value:.1f}s' if is_running else (f'{value:.1f}s' if value else '—')
+        table.add_row(str(idx), u, f'[{style}]{status}[/{style}]', elapsed_str)
+    return table
